+++ src/Cart/resources/config/module.php
@@ -13,6 +13,7 @@ return [
     'preserve_for_user' => false, // Whether to keep and restore user carts across logins and devices
     'merge_duplicates' => false, // Whether to merge carts if `preserve_for_user` is enabled, user logs in and the session contains another cart
     'items' => [
-        'extra_product_attributes_to_merge' => []
+        'extra_product_attributes_to_merge' => [],
+        'shippable_by_default' => null,
     ]
 ];