+++ src/Contracts/CheckoutSubjectItem.php
@@ -16,18 +16,11 @@ namespace Vanilo\Contracts;
 
 interface CheckoutSubjectItem extends Configurable
 {
-    /**
-     * Returns the buyable (product) of the item
-     */
     public function getBuyable(): Buyable;
 
-    /**
-     * Returns the quantity of the line
-     */
     public function getQuantity(): int;
 
-    /**
-     * Returns the (adjusted) line total
-     */
     public function total(): float;
+
+    public function isShippable(): ?bool;
 }