+++ src/Shipment/Tests/ShippingFeeCalculatorTest.php
@@ -14,30 +14,28 @@ declare(strict_types=1);
 
 namespace Vanilo\Shipment\Tests;
 
+use PHPUnit\Framework\Attributes\Test;
 use Vanilo\Shipment\Calculators\NullShippingFeeCalculator;
 use Vanilo\Shipment\Models\ShippingFee;
 use Vanilo\Shipment\Models\ShippingMethod;
 
 class ShippingFeeCalculatorTest extends TestCase
 {
-    /** @test */
-    public function when_having_null_as_calculator_a_null_calculator_object_is_returned()
+    #[Test] public function when_having_null_as_calculator_a_null_calculator_object_is_returned()
     {
         $method = ShippingMethod::create(['name' => 'DPD']);
 
         $this->assertInstanceOf(NullShippingFeeCalculator::class, $method->getCalculator());
     }
 
-    /** @test */
-    public function the_null_calculator_returns_no_adjuster()
+    #[Test] public function the_null_calculator_returns_no_adjuster()
     {
         $method = ShippingMethod::create(['name' => 'Seur']);
 
         $this->assertNull($method->getCalculator()->getAdjuster());
     }
 
-    /** @test */
-    public function the_null_calculator_returns_zero_amount_fee()
+    #[Test] public function the_null_calculator_returns_zero_amount_fee()
     {
         $method = ShippingMethod::create(['name' => 'Seur']);
 
@@ -46,8 +44,7 @@ class ShippingFeeCalculatorTest extends TestCase
         $this->assertEquals(0, $fee->amount()->getValue());
     }
 
-    /** @test */
-    public function the_detailed_amount_can_be_directly_obtained_by_calling_the_estimate_method_of_the_shipping_method()
+    #[Test] public function the_detailed_amount_can_be_directly_obtained_by_calling_the_estimate_method_of_the_shipping_method()
     {
         $method = ShippingMethod::create(['name' => 'Seur']);