+++ src/Shipment/Changelog.md
@@ -9,6 +9,12 @@
 - Changed the minimum Laravel 10 version to v10.48
 - Added Laravel 12 Support
 - Added ETA fields to Shipping Methods
+- Added Shipping Categories
+- Added shipping category and matching condition fields to the shipping methods table
+- BC: Added the following methods to the `ShippingMethod` interface:
+    - `hasShippingCategory()`
+    - `getShippingCategory()`
+    - `getShippingCategoryMatchingCondition()`
 - BC: Added the `getName()` method to the `Carrier` interface
 - Marked the `Carrier::name()` method as deprecated in favor of the `getName()` method